restructured SysInfo class
made the datastructure that holds the system information the outer class
and a builder to gather the information the inner class
